refactor style / structure
