+++ .gitignore
@@ -1,0 +1,5 @@
+venv
+.coverage
+.pytest_cache
+*.egg-info
+.idea
