+++ onto_lib/config.py
@@ -1,12 +1,12 @@
 import pkg_resources as pr
-import os
 from os.path import join
 import json
 
 resource_package = __name__
 OBO_DIR = pr.resource_filename(resource_package, "obo")
-PREFIX_TO_FNAME = pr.resource_filename(resource_package, 
-    "ont_prefix_to_filename.json")
+PREFIX_TO_FNAME = pr.resource_filename(resource_package,
+                                       "ont_prefix_to_filename.json")
+
 
 def ontology_name_to_location():
     prefix_to_location = {}
@@ -14,4 +14,3 @@ def ontology_name_to_location():
         for prefix, fname in json.load(f).items():
             prefix_to_location[prefix] = join(OBO_DIR, fname)
     return prefix_to_location
-
