+++ onto_lib/metadata/term_to_remove_synonyms.json
