+++ setup.cfg
@@ -1,0 +1,6 @@
+[tool:pytest]
+testpaths = tests
+
+[coverage:run]
+branch = True
+source = onto_lib
