+++ setup.py
@@ -1,0 +1,11 @@
+from setuptools import setup, find_packages
+
+# Metadata goes in setup.cfg. These are here for GitHub's dependency graph.
+setup(
+    name="onto_lib",
+    version='1.0.0',
+    packages=find_packages(),
+    include_package_data=True,
+    zip_safe=False,
+    install_requires=[],
+)
